Survival instinct to whales.....
If the Whale detects a positive change in Fullness (the Whale is happy)
-
it lowers the tumbling-rate (less tumbles per second, movement in longer
straight lines).
If the Whale detects a negative change in Fullness (the Whale is sad) -
it raises the tumbling-rate.
If the Whale detects no change in Fullness -
the tumbling rate returns to some default, normal value of tumbling-rate
(this is the adaptation mechanism).